fix: pre-commit

+++ pyhms/sprout/sprout_candidates.py
@@ -1,10 +1,13 @@
 from dataclasses import dataclass, field
 
+import numpy as np
+
 from ..core.individual import Individual
 
 
 @dataclass
 class DemeCandidates:
     """Class for keeping an info about candidate solutions and ELA features data from a single deme."""
+
     individuals: list[Individual]
-    features: dict[str, float | int] = field(default_factory=dict)
+    features: dict[str, np.float64 | np.int64] = field(default_factory=dict)
